# vehicle-management-backend/swagger.json
{
    "swagger" : "2.0",
    "info": {
        "description" : "Documentation of Vehicle Tracking Management API",
        "version": "1.0.0.0",
        "title": "Vehicle Tracking Management APIs",
        "termsOfService" : "http://swagger.io/terms/",
        "contact": {
            "email" : "ciradukundaa@gmail.com",
            "name" : "Clarisse IRADUKUNDA"
        },
        "licence": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
...
    "paths": {
...
        "/carOwner": {
            "post": {
                "tags": [
                    "Car Owner"
                ],
                "summary": "Register car owner",
                "description": "Register Car owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Car Owner object",
                        "required": true,
                        "properties": {
                            "firstName": {
                                "type": "string",
                                "example": "Clarisse"
                            },
                            "lastName": {
                                "type": "string",
                                "example": "Iradukunda"
                            },
                            "nationalID": {
                                "type": "number",
                                "example": 234567
                            },
                            "phone": {
                                "type": "string",
                                "example": "0234567"
                            },
                            "address": {
                                "type": "string",
                                "example": "KG 398 K"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description":"Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "Car Owner"
                ],
                "summary": "Get all car owners",
                "description": "Get all car owners",
                "operationId": "getAllCarOwners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/carOwner/{id}": {
            "get": {
                "tags": [
                    "Car Owner"
                ],
                "summary": "Get car owner by id",
                "description": "Get all car owner details by id",
                "operationId":"getCarOwnerById",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id of car owner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Car Owner"
                ],
                "summary": "Update car owner by Id",
                "description": "update car owner by its id",
                "operationId": "updateCarOwnerById",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Id of the car owner to update",
                        "schema": {
                            "type": "string"
                        },
                        "properties": {
                            "firstName": {
                                "type": "string",
                                "example": "Clarisse"
                            },
                            "lastName": {
                                "type": "string",
                                "example": "Iradukunda"
                            },
                            "nationalID":{
                                "type": "string",
                                "example": "123456789876"
                            },
                            "phone": {
                                "type": "string",
                                "example": "4500000000"
                            },
                            "address": {
                                "type": "string",
                                "example": "KG 398 K"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car owner successfully updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Car Owner"
                ],
                "summary": "Delete car owner by Id",
                "description": "Delete car owner by id",
                "operationId": "deleteCarOwnerById",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id of the car owner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {  
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Car owner not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}